Added some comments to common/util.py

+++ source/test/perf/common/utils.py
@@ -1,3 +1,8 @@
+
+# Utility functions
+
+# Takes a string and return True if it can be converted to a float
+# returns False otherwise 
 def is_float(x):
     try:
         float(x)
@@ -6,6 +11,8 @@ def is_float(x):
     else:
         return True
 
+# Takes a string and return True if it can be converted to an int
+# returns False otherwise 
 def is_int(x):
     try:
         int(x)
@@ -14,12 +21,13 @@ def is_int(x):
     else:
         return True
 
+# Takes a string and return a float if it can be converted to a float
+# returns an int if it can be converted to an int
 def dyn_cast(val):
     if is_float(val):
         return float(val)
     elif is_int(val):
-        return long(val)
+        return int(val)
     else:
         return val
-
